Enhancement: Move Speakers to Past Speakers
