+++ pages/pastconferences/pastspeakers/2021/simon-painter.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Layout from '../../../components/Layout';
-import Header from '../../../components/Header';
-import theme from '../../../theme/theme';
+import Layout from '../../../../components/Layout';
+import Header from '../../../../components/Header';
+import theme from '../../../../theme/theme';
 import Head from 'next/head';
-import SpeakersButton from '../../../components/speakers/SpeakersButton';
+import SpeakersButton from '../../../../components/speakers/SpeakersButton';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import {
   faInstagram,
@@ -14,54 +14,86 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function RizwanaAkmalKhan() {
+export default function SimonPainter() {
   return (
     <Layout>
       <Head>
-        <title>Rizwana Akmal Khan</title>
+        <title>Simon Painter</title>
       </Head>
       <Header
-        title={'Rizwana Akmal Khan'}
+        title={'Simon Painter'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
         <SpeakersButton />
         <img
-          src="/static/speakers/2021/Rizwana_Akmal_Khan.png"
-          alt="Picture of Rizwana Akmal Khan"
+          src="/static/speakers/2021/Simon_Painter.png"
+          alt="Picture of Simon Painter"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> She/Her
+            <b>Pronouns: </b> He/Him
           </ul>
           <ul>
-            <b>Talk Length: </b> 30 mins
+            <b>Talk Length: </b> 60 mins
           </ul>
         </div>
         <br />
-        <h2 tabIndex="0">Good Writers Become Better Developers</h2>
+        <h2 tabIndex="0">SOLID Principles in 5 Nightmares</h2>
         <p tabIndex="0">
-          Good writing might not be rocket science, but it is the
-          secret ingredient in the recipe for a great developer. If
-          you're already an experienced developer or a only just
-          starting on the dev journey, by the end of this talk, you'll
-          come to appreciate the similarities between the two mediums,
-          and be ready to arm yourself with techniques that will help
-          you improve your writing AND development skills.
+          The 5 SOLID principles - popularised by "Uncle" Bob Martin
+          in some of his highly influential books on Object Orientated
+          Software development - are rarely cited directly but they
+          are nevertheless at the heart of a lot of the thinking that
+          goes into modern software development.
+        </p>
+        <p tabIndex="0">
+          These principles have been around in some form or other ever
+          since the 1980s, but continue to be just as relevant today
+          as they were then.
+        </p>
+        <p tabIndex="0">
+          In this talk, we're going to look at each of the 5 SOLID
+          principles, these being:
+        </p>
+        <li>Single Dependency Principle</li>
+        <li>Open/Close Principle</li>
+        <li>Liskov Substitution Principle</li>
+        <li>Interface Segregation Principle</li>
+        <li>Dependency Inversion Principle</li>
+        <p tabIndex="0">See what he did there with the names?</p>
+        <p tabIndex="0">
+          We'll look at each in turn, with the help of some slightly
+          imaginative examples taken from a popular SF franchise. What
+          are they, what nightmare scenarios can occur if they aren't
+          followed, and how they can subsequently be applied.
         </p>
       </section>
       <section>
-        <h2 tabIndex="0">About Rizwana</h2>
+        <h2 tabIndex="0">About Simon</h2>
         <p tabIndex="0">
-          <b>Customer Success Engineer at Vercel</b>
+          <b>.NET Contractor and Consultant</b>
         </p>
         <p tabIndex="0">
-          Rizwana is a designer-developer, poet, podcaster, and
-          occasional-speaker. As a third-culture adult, she has grown
-          up valuing diverse work environments and will always stand
-          in support of spaces where people who may be seen as
-          outsiders are given centre stage.
+          I've been working as a .NET developer for over 13 years now
+          in a variety of industries including government, retail and
+          manufacturing. But as a coder, I've been playing with making
+          computers do whatever my crazed imagination could devise
+          since I was old enough to read my Dad's copy of the ZX
+          Spectrum BASIC coders manual.
+        </p>
+        <p tabIndex="0">
+          I've been speaking about Functional C# at various user
+          groups and conferences around the UK, USA, and India, and am
+          particularly interested in seeing just how far we can push
+          the capabilities of C#.
+        </p>
+        <p tabIndex="0">
+          When I'm not coding, or running after my two small children,
+          I have been known to enjoy the classic series of Doctor Who,
+          Fighting Fantasy Gamebooks, Cryptic Crosswords, and rather
+          more coffee than is probably good for me.
         </p>
       </section>
 
@@ -71,30 +103,10 @@ export default function RizwanaAkmalKhan() {
           <a
             target="_blank"
             rel="noreferrer"
-            aria-label="Rizwana Akmal Khan: Twitter"
-            href="https://twitter.com/rizbizkits"
+            aria-label="Simon Painter: Twitter"
+            href="https://twitter.com/madSimonJ"
           >
             Twitter
-          </a>
-        </li>
-        <li>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Rizwana Akmal Khan: Website"
-            href="https://rizwanakhan.com/"
-          >
-            Website
-          </a>
-        </li>
-        <li>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            aria-label="Rizwana Akmal Khan: LinkedIn"
-            href="https://www.linkedin.com/in/riz/"
-          >
-            LinkedIn
           </a>
         </li>
       </section>
