+++ pages/pastconferences/pastspeakers/pastspeakergrid.js
@@ -22,6 +22,121 @@ export default function Speaker2021Component() {
         </p>
       </section>
       <section>
+        <h2>2021</h2>
+      </section>
+      <section className="speaker-grid">
+        <SpeakerGridItemComponent
+          speakername="Tania Allard"
+          pic="/static/speakers/2021/Tania_Allard.jpg"
+          link="../2021/tania-allard"
+          talkbrief="KEYNOTE: Open source for fun and for profit"
+        />
+      </section>
+      <section className="speaker-grid">
+        <SpeakerGridItemComponent
+          speakername="Lex Lofthouse"
+          pic="/static/speakers/2021/Lex_Lofthouse.png"
+          link="../2021/lex-lofthouse"
+          talkbrief="Design for Developers"
+        />
+        <SpeakerGridItemComponent
+          speakername="Connell"
+          pic="/static/speakers/2021/Connell_W.png"
+          link="../2021/connell-w"
+          talkbrief="Onion Architecture with DDD and CQRS"
+        />
+        <SpeakerGridItemComponent
+          speakername="Dylan Beattie"
+          pic="/static/speakers/2021/Dylan_Beattie.jpg"
+          link="../2021/dylan-beattie"
+          talkbrief="There's No Such Thing As Plain Text"
+        />
+        <SpeakerGridItemComponent
+          speakername="Simon Painter"
+          pic="/static/speakers/2021/Simon_Painter.png"
+          link="../2021/simon-painter"
+          talkbrief="SOLID Principles in 5 Nightmares"
+        />
+        <SpeakerGridItemComponent
+          speakername="Steve Collins"
+          pic="/static/speakers/2021/Steve_Collins.jpg"
+          link="../2021/steve-collins"
+          talkbrief="The [Source Code] Generation Game"
+        />
+        <SpeakerGridItemComponent
+          speakername="Stephen Haunts"
+          pic="/static/speakers/2021/Stephen_Haunts.jpg"
+          link="../2021/stephen-haunts"
+          talkbrief="Hacking Humans"
+        />
+        <SpeakerGridItemComponent
+          speakername="Manu Magalhaes"
+          pic="/static/speakers/2021/Manu_Magalhaes.png"
+          link="../2021/manu-magalhaes"
+          talkbrief="Nitty-Gitty: Master Git from the Inside"
+        />
+        <SpeakerGridItemComponent
+          speakername="James Croft"
+          pic="/static/speakers/2021/James_Croft.png"
+          link="../2021/james-croft"
+          talkbrief="Should you write UI tests?"
+        />
+        <SpeakerGridItemComponent
+          speakername="Matt Whetton"
+          pic="/static/speakers/2021/Matt_Whetton.png"
+          link="../2021/matt-whetton"
+          talkbrief="We need to talk about technical debt"
+        />
+        <SpeakerGridItemComponent
+          speakername="Clifford Agius"
+          pic="/static/speakers/2021/Clifford_Agius.jpg"
+          link="../2021/clifford-agius"
+          talkbrief="3D printed Bionic Hand a little IOT and a Xamarin Mobile App"
+        />
+        <SpeakerGridItemComponent
+          speakername="Poornima Nayar"
+          pic="/static/speakers/2021/Poornima_Nayar.png"
+          link="../2021/poornima-nayar"
+          talkbrief="REST, GraphQL and gRPC : A Comparison"
+        />
+        <SpeakerGridItemComponent
+          speakername="Tom Morton"
+          pic="/static/speakers/2021/Tom_Morton.png"
+          link="../2021/tom-morton"
+          talkbrief="Open Banking vs. DeFi; The technological future of finance?"
+        />
+        <SpeakerGridItemComponent
+          speakername="Kojo Hinson"
+          pic="/static/speakers/2021/Kojo_Hinson.png"
+          link="../2021/kojo-hinson"
+          talkbrief="Embracing Skynet - An exploration of GPT-3 and its potential applications"
+        />
+        <SpeakerGridItemComponent
+          speakername="Stephen Jackson"
+          pic="/static/speakers/2021/Stephen_Jackson.png"
+          link="../2021/stephen-jackson"
+          talkbrief="Senior By Default"
+        />
+        <SpeakerGridItemComponent
+          speakername="Rizwana Akmal Khan"
+          pic="/static/speakers/2021/Rizwana_Akmal_Khan.png"
+          link="../2021/riz-akmal-khan"
+          talkbrief="Good Writers Become Better Developers"
+        />
+        <SpeakerGridItemComponent
+          speakername="Leke Sholuade"
+          pic="/static/speakers/2021/Leke_Sholuade.jpg"
+          link="../2021/leke-sholuade"
+          talkbrief="Why do we need a Black Valley"
+        />
+        <SpeakerGridItemComponent
+          speakername="Jennifer Mackown"
+          pic="/static/speakers/2021/Jen_Mackown.png"
+          link="../2021/jennifer-mackown"
+          talkbrief="How to ruin kid's games with machine learning"
+        />
+      </section>
+      <section>
         <h2>2019</h2>
       </section>
       <section className="speaker-grid">
